Implemented comments
