--- RUFAS/biophysical/animal/data_types/growth_outputs.py
@@ -1,10 +1,0 @@
-from dataclasses import dataclass
-
-from RUFAS.routines.animal.life_cycle.animal_events import AnimalEvents
-
-
-@dataclass
-class GrowthOutputs:
-    body_weight: float
-    conceptus_weight: float
-    events: AnimalEvents
