+++ RUFAS/biophysical/animal/digestive_system/digestive_system.py
@@ -1,7 +1,7 @@
 from RUFAS.biophysical.animal.animal_config import AnimalConfig
 from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.biophysical.animal.data_types.animal_types import AnimalType
-from RUFAS.biophysical.animal.data_types.digestive_system_inputs import DigestiveSystemInputs
+from RUFAS.biophysical.animal.data_types.digestive_system import DigestiveSystemInputs
 from RUFAS.biophysical.animal.digestive_system.enteric_methane_calculator import EntericMethaneCalculator
 from RUFAS.biophysical.animal.digestive_system.manure_excretion_calculator import ManureExcretionCalculator
 
@@ -14,32 +14,29 @@ class DigestiveSystem:
     """
     manure_excretion: AnimalManureExcretions
     phosphorus_excreted: float
-    methane_emission: float
+    enteric_methane_emission: float
 
     def __init__(self) -> None:
         self.manure_excretion = AnimalManureExcretions()
         self.phosphorus_excreted = 0.0
-        self.methane_emission = 0.0
+        self.enteric_methane_emission = 0.0
 
     def process_digestion(self, digestive_system_inputs: DigestiveSystemInputs) -> None:
         """
-        Handles an animal's daily digest updates.
+        Processes the digestion for different types of animals by calculating methane emission
+        and manure excretion based on the provided digestive system inputs.
 
         Parameters
         ----------
-        general_properties: GeneralProperties
-            Animal properties that are general or are used to determine many animal outcomes.
-        animal_nutrient_property: AnimalGrowthProperties
-            Animal properties that are related to animal nutrients.
-        milk_production_properties: MilkProductionProperties
-            Animal properties that are related to animal milk production.
+        digestive_system_inputs : DigestiveSystemInputs
+            Contains inputs related to the digestive system of the animal, including animal type,
+            body weight, nutrient details, fecal phosphorus, and urine phosphorus requirements.
 
-        Returns
-        -------
-        tuple[dict[str, float], AnimalManureExcretions]
-            A dictionary that contains the manure excretion values as specified
-            in the AnimalManureExcretions class definition.
-
+        Raises
+        ------
+        TypeError
+            If the animal type in digestive_system_inputs is not supported, a TypeError is raised
+            with information about supported animal types.
         """
         om = OutputManager()
         if digestive_system_inputs.animal_type == AnimalType.CALF:
@@ -54,7 +51,7 @@ class DigestiveSystem:
                 digestive_system_inputs.nutrients,
                 digestive_system_inputs.nutrient_concentrations,
             )
-            self.methane_emission = methane_emission
+            self.enteric_methane_emission = methane_emission
             self.phosphorus_excreted = phosphorus
             self.manure_excretion = excretion
             return
@@ -73,7 +70,7 @@ class DigestiveSystem:
                 digestive_system_inputs.nutrients,
                 digestive_system_inputs.nutrient_concentrations,
             )
-            self.methane_emission = methane_emission
+            self.enteric_methane_emission = methane_emission
             self.phosphorus_excreted = phosphorus
             self.manure_excretion = excretion
             return
@@ -103,7 +100,7 @@ class DigestiveSystem:
                 digestive_system_inputs.nutrient_concentrations,
             )
 
-            self.methane_emission = methane_emission
+            self.enteric_methane_emission = methane_emission
             self.phosphorus_excreted = phosphorus
             self.manure_excretion = excretion
             return
